Move all writing tests into tests/
